Add scale-in cooldown support
This reimplements ASG scale-in cooldown inside the lambda. It takes two
parameters which correspond to the existing ASG parameters in the
elastic CI stack.

1. `SCALE_IN_COOLDOWN_PERIOD` is the cooldown time between scale in
   events. This defaults to the existing 5 minutes.

2. `SCALE_IN_ADJUSTMENT` is the maximum adjustment during scale-in
   events. Unlike the ASG we may scale in less if we calculate that the
   desired is closer than the adjustment. This defaults to the existing
   -1.

This cheats a bit by storing `lastScaleInTime` in a global variable.
This means we'll forget about our cooldown during a cold start. This
should happen fairly infrequently and just make us a bit aggressive
about scaling it; it shouldn't affect correctness.