run deploy steps on the new "deploy"  queue
The "deploy-legacy" queue is deprecated, and we're slowly moving steps
off it.

The agents processing the "deploy" queue are an elastic stack, and the
awscli is available. It's likely we'll need to update some IAM roles
used by those agents, but we do that elsewhere.

One of the steps also uses the github-release tool, which isn't
available on the agents. We can use the deploytools image (via a
buildkite plugin) to get it though.

github-release also requires a github API token. On the old agents that
was set in an agent hook, but now we need to read it from SSM.