Only search scale in and scale out if they are not respectively disabled
If any one of them were disabled, then the corresponding scaling events
would not appear in the scaling activities and so the entire scaling
activity history would have to be searched each time.

In the elastic ci stack for aws, this was the case as scale in was
disabled. Many customers have reported that there were excessive calls
and rate limiting of the DescribeScalingActivites endpoint, so this was a real problem.

Also there is no need to check environment variables on each iteration
of the main loop, so we have moved them outside.